+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.2-alpha.1] - 2023-04-27
+
+### Added
+
+- Added Toolbar method to create set of default methods.
+- Added Toolbar method to replace Packages Manifest file with a clean preset file.
+- Added Toolbar method to setup a new project combining the above methods.
 
 ## [0.2.1] - 2023-01-24